requirements.txt
pandas
numpy
scikit-learn
pydicom
highdicom
tqdm
PyYAML
transformers
torch
scipy
picai_prep 